extraction of mocked classes reverted
